Il2CppDumper Wrapper (GUI)
I created this tool, for users who don't want to work with terminal 24/7

It works!

To be honest I used AI for the code-base.

But I had to make some changes to fix the damn issues it comes with it. & Personal changes.

It should also proof how our Il2CppDumper could be integrated with Wrappers